Fix types of constant

+++ Include/GWCA/Constants/Constants.h
@@ -182,10 +182,10 @@ namespace GW {
 
         enum class InterfaceSize { SMALL = 0, NORMAL, LARGE, LARGER };
         namespace HealthbarHeight {
-            constexpr int Small = 25;
-            constexpr int Normal = 22;
-            constexpr int Large = 28;
-            constexpr int Larger = 31;
+            constexpr size_t Small = 25;
+            constexpr size_t Normal = 22;
+            constexpr size_t Large = 28;
+            constexpr size_t Larger = 31;
         }
 
         // travel, region, districts
